spunoff config
