+++ config.mjs
@@ -1,0 +1,9 @@
+const config={}
+config.state=
+{
+	palette:['#000'],
+	pointers:{},
+	pts:{},
+	viewbox:{height:150,width:300,x:0,y:0}
+}
+export default config
